handpose/data_preparation/requirement.txt
tqdm
regex
projectaria-tools==1.0.0
opencv-python==4.7.0.72
av==10.0.0
numpy==1.24.2
pandas==1.3.5
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
black==22.3.0
ufmt==2.0.0
usort==1.0.3